[BEAM-6030] Split metrics related options out of PipelineOptions
