+++ runners/core-java/src/test/java/org/apache/beam/runners/core/metrics/MetricsPusherTest.java
@@ -23,7 +23,7 @@ import static org.junit.Assert.assertThat;
 import org.apache.beam.sdk.io.GenerateSequence;
 import org.apache.beam.sdk.metrics.Counter;
 import org.apache.beam.sdk.metrics.Metrics;
-import org.apache.beam.sdk.options.PipelineOptions;
+import org.apache.beam.sdk.metrics.MetricsOptions;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.testing.UsesAttemptedMetrics;
 import org.apache.beam.sdk.testing.UsesCounterMetrics;
@@ -53,7 +53,7 @@ public class MetricsPusherTest {
   @Before
   public void init() {
     TestMetricsSink.clear();
-    PipelineOptions options = pipeline.getOptions();
+    MetricsOptions options = pipeline.getOptions().as(MetricsOptions.class);
     options.setMetricsSink(TestMetricsSink.class);
   }
 
@@ -67,7 +67,8 @@ public class MetricsPusherTest {
         .apply(ParDo.of(new CountingDoFn()));
     pipeline.run();
     // give metrics pusher time to push
-    Thread.sleep((pipeline.getOptions().getMetricsPushPeriod() + 1L) * 1000);
+    Thread.sleep(
+        (pipeline.getOptions().as(MetricsOptions.class).getMetricsPushPeriod() + 1L) * 1000);
     assertThat(TestMetricsSink.getCounterValue(), is(NUM_ELEMENTS));
   }
