+++ runners/extensions-java/metrics/src/test/java/org/apache/beam/runners/extensions/metrics/MetricsGraphiteSinkTest.java
@@ -23,7 +23,7 @@ import java.io.IOException;
 import java.net.ServerSocket;
 import java.util.concurrent.CountDownLatch;
 import org.apache.beam.sdk.metrics.MetricQueryResults;
-import org.apache.beam.sdk.options.PipelineOptions;
+import org.apache.beam.sdk.metrics.MetricsOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -59,7 +59,7 @@ public class MetricsGraphiteSinkTest {
   @Test
   public void testWriteMetricsWithCommittedSupported() throws Exception {
     MetricQueryResults metricQueryResults = new CustomMetricQueryResults(true);
-    PipelineOptions pipelineOptions = PipelineOptionsFactory.create();
+    MetricsOptions pipelineOptions = PipelineOptionsFactory.create().as(MetricsOptions.class);
     pipelineOptions.setMetricsGraphitePort(port);
     pipelineOptions.setMetricsGraphiteHost("127.0.0.1");
     MetricsGraphiteSink metricsGraphiteSink = new MetricsGraphiteSink(pipelineOptions);
@@ -89,7 +89,7 @@ public class MetricsGraphiteSinkTest {
   @Test
   public void testWriteMetricsWithCommittedUnSupported() throws Exception {
     MetricQueryResults metricQueryResults = new CustomMetricQueryResults(false);
-    PipelineOptions pipelineOptions = PipelineOptionsFactory.create();
+    MetricsOptions pipelineOptions = PipelineOptionsFactory.create().as(MetricsOptions.class);
     pipelineOptions.setMetricsGraphitePort(port);
     pipelineOptions.setMetricsGraphiteHost("127.0.0.1");
     MetricsGraphiteSink metricsGraphiteSink = new MetricsGraphiteSink(pipelineOptions);
