+++ runners/spark/src/test/java/org/apache/beam/runners/spark/metrics/SparkMetricsPusherTest.java
@@ -28,7 +28,7 @@ import org.apache.beam.runners.spark.io.CreateStream;
 import org.apache.beam.sdk.coders.VarIntCoder;
 import org.apache.beam.sdk.metrics.Counter;
 import org.apache.beam.sdk.metrics.Metrics;
-import org.apache.beam.sdk.options.PipelineOptions;
+import org.apache.beam.sdk.metrics.MetricsOptions;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.testing.ValidatesRunner;
 import org.apache.beam.sdk.transforms.Create;
@@ -63,7 +63,7 @@ public class SparkMetricsPusherTest {
   @Before
   public void init() {
     TestMetricsSink.clear();
-    PipelineOptions options = pipeline.getOptions();
+    MetricsOptions options = pipeline.getOptions().as(MetricsOptions.class);
     options.setMetricsSink(TestMetricsSink.class);
   }
 
@@ -94,7 +94,8 @@ public class SparkMetricsPusherTest {
 
     pipeline.run();
     // give metrics pusher time to push
-    Thread.sleep((pipeline.getOptions().getMetricsPushPeriod() + 1L) * 1000);
+    Thread.sleep(
+        (pipeline.getOptions().as(MetricsOptions.class).getMetricsPushPeriod() + 1L) * 1000);
     assertThat(TestMetricsSink.getCounterValue(), is(6L));
   }
 
@@ -119,7 +120,8 @@ public class SparkMetricsPusherTest {
 
     pipeline.run();
     // give metrics pusher time to push
-    Thread.sleep((pipeline.getOptions().getMetricsPushPeriod() + 1L) * 1000);
+    Thread.sleep(
+        (pipeline.getOptions().as(MetricsOptions.class).getMetricsPushPeriod() + 1L) * 1000);
     assertThat(TestMetricsSink.getCounterValue(), is(6L));
   }
 }
