+++ sdks/java/core/src/main/java/org/apache/beam/sdk/options/PipelineOptions.java
@@ -31,7 +31,6 @@ import java.util.concurrent.atomic.AtomicLong;
 import javax.annotation.concurrent.ThreadSafe;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.PipelineRunner;
-import org.apache.beam.sdk.metrics.MetricsSink;
 import org.apache.beam.sdk.options.ProxyInvocationHandler.Deserializer;
 import org.apache.beam.sdk.options.ProxyInvocationHandler.Serializer;
 import org.apache.beam.sdk.transforms.DoFn;
@@ -402,58 +401,4 @@ public interface PipelineOptions extends HasDisplayData {
       return String.format("%s/%s", info.getName(), info.getVersion()).replace(" ", "_");
     }
   }
-
-  @Description("The beam sink class to which the metrics will be pushed")
-  @Default.InstanceFactory(NoOpMetricsSink.class)
-  Class<? extends MetricsSink> getMetricsSink();
-
-  void setMetricsSink(Class<? extends MetricsSink> metricsSink);
-
-  /**
-   * A {@link DefaultValueFactory} that obtains the class of the {@code NoOpMetricsSink} if it
-   * exists on the classpath, and throws an exception otherwise.
-   *
-   * <p>As the {@code NoOpMetricsSink} is in an independent module, it cannot be directly referenced
-   * as the {@link Default}. However, it should still be used if available.
-   */
-  class NoOpMetricsSink implements DefaultValueFactory<Class<? extends MetricsSink>> {
-    @Override
-    public Class<? extends MetricsSink> create(PipelineOptions options) {
-      try {
-        @SuppressWarnings({"unchecked", "rawtypes"})
-        Class<? extends MetricsSink> noOpMetricsSinkClass =
-            (Class<? extends MetricsSink>)
-                Class.forName(
-                    "org.apache.beam.runners.core.metrics.NoOpMetricsSink",
-                    true,
-                    ReflectHelpers.findClassLoader());
-        return noOpMetricsSinkClass;
-      } catch (ClassNotFoundException e) {
-        throw new IllegalArgumentException(
-            String.format("NoOpMetricsSink was not found on classpath"));
-      }
-    }
-  }
-
-  @Description("The metrics push period in seconds")
-  @Default.Long(5)
-  Long getMetricsPushPeriod();
-
-  void setMetricsPushPeriod(Long period);
-
-  @Description("MetricsHttpSink url")
-  String getMetricsHttpSinkUrl();
-
-  void setMetricsHttpSinkUrl(String metricsSink);
-
-  @Description("The graphite metrics host")
-  String getMetricsGraphiteHost();
-
-  void setMetricsGraphiteHost(String host);
-
-  @Description("The graphite metrics port")
-  @Default.Integer(2003)
-  Integer getMetricsGraphitePort();
-
-  void setMetricsGraphitePort(Integer port);
 }
